Added Copy/Paste and select features to table in right of way tab
Updated right of way tab to include Copy/Paste features.  Can copy/paste
back to and from Excel.  Data is copied to clipboard as tab and newline
delimited text.  If data in clipboard is erroneous it will alert in the
status bar.  Data range checking occurs for updated table cell as
normal.