Added export results feature.  Added diagnostics tab for display of calculation matrices.  Updated LowFiTable.fill_table()
Added export feature to Results tab.  Results table can be written to
.csv file.
Added a Diagnostics tab.  Scalars, 1D and 2D lists and Numpy arrays can
be logged to this tab.  Tables on this tab can be copied to clipboard
and exported to .csv file (as a batch. files named as per logged
titles).  Shortcuts don't work on Diagnostics screen at this stage,
right click copy will work.
Added methods to main to support output of matrices to diagnostics tab.
Updated LowFiTable.fill_table() to correctly display some numpy arrays
that were being filled as a string of a list.  Included an option to
print complex numbers in the form 0, 1j or 1+2j rather than the default
(0j) or (1+0j).